refactor: refactor of theses and supervisors listing
