+++ frontend/next.config.ts
@@ -1,9 +1,9 @@
-import type { NextConfig } from "next";
+import type { NextConfig } from "next"
 
 const nextConfig: NextConfig = {
   env: {
     ITEMS_PER_PAGE: process.env.ITEMS_PER_PAGE,
   },
-};
+}
 
-export default nextConfig;
+export default nextConfig
